Half-baked update to the rendering: Try to do more in mm units rather than depending on pixels.
Reduce the frame thickness. Uses rounded corners. Draws frame on top
of text and image. Still work to do on the text alignment and scale.

Another note: OpenVG doesn't implement a view matrix stack à la
OpenGL. So, we probably have to implement our own (as the manual says)
in order to do more drawing in mm.