Add support for the pibrella's big red button.
For now, pressing the button exits the program (something
new!). Eventually, maybe we'll try to play/pause MPD, but that
involves keeping some state of the player.

The primary point here is that I discovered it was leaking
memory. After running for a couple of weeks, the OOM killer offed the
process. But, I can't really figure out what's happening because:
Valgrind won't work on the Pi.